Deprecate PostgreSQL 9.5. Eliminate most references and reduce visibility of unsupported versions.
