+++ core/src/org/labkey/core/dialect/PostgreSql92Dialect.java
@@ -48,7 +48,7 @@ import java.util.regex.Pattern;
  * Date: 5/21/12
  * Time: 8:52 AM
  */
-public abstract class PostgreSql92Dialect extends PostgreSql91Dialect
+abstract class PostgreSql92Dialect extends PostgreSql91Dialect
 {
     protected PostgreSql92Dialect()
     {
@@ -126,49 +126,4 @@ public abstract class PostgreSql92Dialect extends PostgreSql91Dialect
         _inClauseGenerator = getJdbcVersion(scope) >= 4 ? new ArrayParameterInClauseGenerator(scope) : new ParameterMarkerInClauseGenerator();
     }
 
-    /*
-     TestCase migrated from PostgreSql91Dialect when that class promoted to api.
-     */
-    public static class TestCase extends Assert
-    {
-        PostgreSql92Dialect getDialect()
-        {
-            DbSchema core = CoreSchema.getInstance().getSchema();
-            SqlDialect d = core.getSqlDialect();
-            if (d instanceof PostgreSql92Dialect)
-                return (PostgreSql92Dialect)d;
-            return null;
-        }
-
-        @Test
-        public void testInClause()
-        {
-            PostgreSql92Dialect d = getDialect();
-            if (null == d)
-                return;
-            DbSchema core = CoreSchema.getInstance().getSchema();
-
-            SQLFragment shortSql = new SQLFragment("SELECT COUNT(*) FROM core.usersdata WHERE userid ");
-            d.appendInClauseSql(shortSql, Arrays.asList(1, 2, 3));
-            assertEquals(1, new SqlSelector(core, shortSql).getRowCount());
-
-            ArrayList<Object> l = new ArrayList<>();
-            for (int i=1 ; i<=TEMPTABLE_GENERATOR_MINSIZE+1 ; i++)
-                l.add(i);
-            SQLFragment longSql = new SQLFragment("SELECT COUNT(*) FROM core.usersdata WHERE userid ");
-            d.appendInClauseSql(longSql, l);
-            assertEquals(1, new SqlSelector(core, longSql).getRowCount());
-
-            SQLFragment shortSqlStr = new SQLFragment("SELECT COUNT(*) FROM core.usersdata WHERE displayname ");
-            d.appendInClauseSql(shortSqlStr, Arrays.asList("1", "2", "3"));
-            assertEquals(1, new SqlSelector(core, shortSqlStr).getRowCount());
-
-            l = new ArrayList<>();
-            for (int i=1 ; i<=TEMPTABLE_GENERATOR_MINSIZE+1 ; i++)
-                l.add(String.valueOf(i));
-            SQLFragment longSqlStr = new SQLFragment("SELECT COUNT(*) FROM core.usersdata WHERE displayname ");
-            d.appendInClauseSql(longSqlStr, l);
-            assertEquals(1, new SqlSelector(core, longSqlStr).getRowCount());
-        }
-    }
 }
